Further testing. Start refactoring to run locally.
