+++ package.json
@@ -9,7 +9,8 @@
   "scripts": {
     "build": "node make.js build",
     "test": "node make.js test",
-    "testLegacy": "node make.js testLegacy"
+    "testLegacy": "node make.js testLegacy",
+    "layout": "node make.js layout"
   },
   "repository": {
     "type": "git",
